+++ Stack/cpluscplus/Stack.cpp
@@ -11,7 +11,7 @@ public:
     virtual void clear() = 0;
     virtual void push(T) = 0;
     virtual T pop() = 0;
-    virtual int length() = 0;
+    virtual long long length() = 0;
     virtual T top() = 0;
     virtual void setDirection(int) {} //only Array base implementation will override setDirection as this is not relevant for LinkList based implementatin.
 };
